\name{varhandle-package}
\alias{varhandle-package}
\alias{varhandle}
\docType{package}
\title{
    A Package to Work and Handl Variables Easier, Convenient and Flawless
}
\description{
    A collection of functions for working and handling variables easier, convenient and flawless in R programming environment. It is suitable for Projects: Small to semi-large projects with massive amount of data. Programmers: Programming and exploring data on the fly or preparing and assembling a pipeline.
}
\details{
    \tabular{ll}{
    Package: \tab varhandle\cr
    Type: \tab Package\cr
    Version: \tab 2.0\cr
    Date: \tab 2016-04-14\cr
    License: \tab GPL (>=2)\cr
    }
    ~~ An overview of how to use the package, including the most important ~~
    ~~ functions ~~
}
\author{
    Mehrad Mahmoudian [aut, cre]
    Who to complain to <m.mahmoudian@gmail.com>
        Reporting issues: https://bitbucket.org/mehrad_mahmoudian/varhandle/issues?status=new&status=open
}
%\references{
%    ~~ Literature or other references for background information ~~
%}

\keyword{ package }

%%\seealso{

%%}
%%\examples{
%%    ~~ simple examples of the most important functions ~~
%%}
